requirements.txt
# add the file location where the TensorFlow library is saved e.g file:///Users/username/Desktop/project_folder/TensorFlow/models/research
# and uncomment the following line: 
# object-detection @ file:///your_path/models/research

absl-py==0.12.0
apache-beam==2.33.0
astunparse==1.6.3
attrs==21.2.0
avro-python3==1.9.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.24
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
ensemble-boxes==1.0.7
fastavro==1.4.7
Flask==2.0.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-api-core==2.2.1
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kaggle==1.5.12
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.37.0
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
numba==0.54.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
orjson==3.6.4
pandas==1.3.4
Pillow==8.4.0
portalocker==2.3.2
promise==2.3
protobuf==3.19.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydot==1.4.2
pymongo==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==2.0.0
scikit-learn==1.0.1
scipy==1.7.1
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-io==0.21.0
tensorflow-io-gcs-filesystem==0.21.0
tensorflow-metadata==1.4.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.6.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.6.0
tf-slim==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
typeguard==2.13.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
